requirements.txt
# =============================================================================
# Multi-Agent Research System - Python Dependencies
# =============================================================================

# Core Pydantic AI with multiple provider support
pydantic-ai[openai,anthropic,gemini]>=0.0.14

# Configuration management
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP client for async API calls
httpx>=0.27.0

# Google API clients for Gmail integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# CLI interface and formatting
rich>=13.0.0

# Testing framework
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Code quality tools
ruff>=0.1.0
mypy>=1.7.0

# Type stubs for Google libraries
types-google-api-python-client>=1.0.0
